+++ source/modulo_components/test/python/test_component_communication.py
@@ -2,17 +2,26 @@ import pytest
 from modulo_components.component import Component
 
 
+class Trigger(Component):
+    def __init__(self):
+        super().__init__("trigger")
+        self.add_trigger("test")
+
+    def trigger(self):
+        super().trigger("test")
+
+
 @pytest.mark.parametrize("minimal_cartesian_input", [[Component, "/topic"]], indirect=True)
 @pytest.mark.parametrize("minimal_cartesian_output", [[Component, "/topic"]], indirect=True)
-def test_input_output(ros_exec, random_state, minimal_cartesian_output, minimal_cartesian_input):
+def test_input_output(ros_exec, random_pose, minimal_cartesian_output, minimal_cartesian_input):
     ros_exec.add_node(minimal_cartesian_input)
     ros_exec.add_node(minimal_cartesian_output)
     ros_exec.spin_until_future_complete(minimal_cartesian_input.received_future, timeout_sec=0.5)
     ros_exec.remove_node(minimal_cartesian_input)
     ros_exec.remove_node(minimal_cartesian_output)
     assert minimal_cartesian_input.received_future.result()
-    assert random_state.get_name() == minimal_cartesian_input.input.get_name()
-    assert random_state.dist(minimal_cartesian_input.input) < 1e-3
+    assert random_pose.get_name() == minimal_cartesian_input.input.get_name()
+    assert random_pose.dist(minimal_cartesian_input.input) < 1e-3
 
 
 @pytest.mark.parametrize("minimal_cartesian_input", [[Component, "/topic"]], indirect=True)
@@ -22,3 +31,17 @@ def test_input_output_invalid(ros_exec, make_minimal_invalid_encoded_state_publi
     ros_exec.add_node(invalid_publisher)
     ros_exec.spin_until_future_complete(minimal_cartesian_input.received_future, timeout_sec=0.5)
     assert not minimal_cartesian_input.received_future.result()
+
+
+def test_trigger(ros_exec, make_predicates_listener):
+    trigger = Trigger()
+    listener = make_predicates_listener("trigger", ["test"])
+    ros_exec.add_node(listener)
+    ros_exec.add_node(trigger)
+    ros_exec.spin_until_future_complete(listener.predicates_future, timeout_sec=0.5)
+    assert not listener.predicates_future.done()
+    assert not listener.predicate_values["test"]
+    trigger.trigger()
+    ros_exec.spin_until_future_complete(listener.predicates_future, timeout_sec=0.5)
+    assert listener.predicates_future.done()
+    assert listener.predicate_values["test"]